requirements.txt

# Core Dependencies
ollama>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async & Networking
aiohttp>=3.9.0
asyncio>=3.4.3
httpx>=0.25.0
duckduckgo-search>=5.0.0
playwright>=1.40.0

# Vector Database & Memory
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Caching & Queue
redis>=5.0.1
celery>=5.3.4

# CLI & UI
rich>=13.7.0
typer>=0.9.0
streamlit>=1.29.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0.1
loguru>=0.7.2
tenacity>=8.2.3
jinja2>=3.1.2
beautifulsoup4>=4.12.0
pypdf>=3.17.0
python-docx>=1.1.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.1
ruff>=0.1.9
